backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
supabase==2.7.4
httpx==0.27.2
PyJWT==2.9.0
python-multipart==0.0.9
reportlab==4.2.2
Pillow==10.4.0
python-dotenv==1.0.1
